+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 **3.4.1** (in progress, updated 2 May 2021):
 
-*Pre-emptive version bump, no user-visible changes yet.*
+**Changed**:
+
+- Update docs: as of `unpythonic` 0.15, it runs on `mcpyrate`, and provides fully functional example dialects based on a whole-module AST transformation.
 
 
 ---